--- a/lineage_test.go
+++ b/lineage_test.go
@@ -46,11 +46,14 @@ func TestBindLineage(t *testing.T) {
 		fmt.Fprintf(tc, "Schema versions: %s\n", lin.allVersions())
 
 		// lenseselem := cue.Hid("_sortedLenses", "github.com/grafana/thema")
-		lenseselem := cue.Str("lenses")
-		slpath := cue.MakePath(lenseselem)
-		llen, err := lin.Underlying().LookupPath(slpath).Len().Int64()
-		if err != nil {
-			tc.Fatal("error getting schemas len", err)
+		var llen int64
+		if slen > 1 {
+			lenseselem := cue.Str("lenses")
+			slpath := cue.MakePath(lenseselem)
+			llen, err = lin.Underlying().LookupPath(slpath).Len().Int64()
+			if err != nil {
+				tc.Fatal("error getting lenses len", err)
+			}
 		}
 		fmt.Fprintf(tc, "Lenses count: %v\n", llen)
 	})