--- a/testdata/invalidlineage/empty-schemas.txtar
+++ b/testdata/invalidlineage/empty-schemas.txtar
@@ -7,7 +7,3 @@ name: "empty-schemas"
 schemas: []
 -- out/bindfail --
 incompatible list lengths (0 and 1)
-_sortedSchemas: invalid value [] (does not satisfy list.MinItems(1)): len(list) < MinItems(1) (0 < 1):
-    ../../lineage.cue:98:7
-    ../../lineage.cue:94:21
-    ../../lineage.cue:98:21