--- a/testdata/invalidlineage/second-schema-versionless.txtar
+++ b/testdata/invalidlineage/second-schema-versionless.txtar
@@ -32,8 +32,8 @@ lenses: [
 ]
 -- out/bindfail --
 _schemasAreOrdered.1: conflicting values true and false:
-    ../../lineage.cue:126:3
-    ../../lineage.cue:127:4
-    ../../lineage.cue:132:5
-    ../../lineage.cue:133:11
+    ../../lineage.cue:113:3
+    ../../lineage.cue:114:4
+    ../../lineage.cue:119:5
+    ../../lineage.cue:120:11
     ./in.cue:4:1