--- a/testdata/isappendonly/invalid/default.txtar
+++ b/testdata/isappendonly/invalid/default.txtar
@@ -22,6 +22,6 @@ lin2: schemas: [{
 }]
 -- out/isappendonly-fail --
 field aunion not present in {aunion:*"bar" | "foo" | "baz"}:
-    ../../../lineage.cue:248:10
+    ../../../lineage.cue:235:10
     ./in.cue:16:13
 missing field "aunion"