--- a/testdata/isappendonly/invalid/noref.txtar
+++ b/testdata/isappendonly/invalid/noref.txtar
@@ -34,7 +34,7 @@ lin2: schemas: [{
 }]
 -- out/isappendonly-fail --
 field #Baz not present in {someField:string,#Baz:{run:string,tell:bytes}}:
-    ../../../lineage.cue:247:10
+    ../../../lineage.cue:248:10
     ./in.cue:22:13
 missing field "#Baz"
 required field is optional in subsumed value: dat