--- a/testdata/isappendonly/invalid/optional.txtar
+++ b/testdata/isappendonly/invalid/optional.txtar
@@ -22,5 +22,7 @@ lin2: schemas: [{
     }
 }]
 -- out/isappendonly-fail --
-field not allowed in closed struct: aNewOptionalField
-value not an instance
+field aNewOptionalField not present in {aField:string}:
+    ../../../lineage.cue:248:10
+    ./in.cue:7:10
+missing field "aNewOptionalField"