--- a/testdata/isappendonly/invalid/stringconstraint.txtar
+++ b/testdata/isappendonly/invalid/stringconstraint.txtar
@@ -35,7 +35,7 @@ lin2: schemas: [{
 }]
 -- out/isappendonly-fail --
 field aString not present in {aString:strings.MinRunes(2),anObject:{aField:int}}:
-    ../../../lineage.cue:248:10
+    ../../../lineage.cue:235:10
     ./in.cue:25:10
 missing field "aString"
 invalid value strings.MinRunes(2) (does not satisfy strings.MinRunes(1)): error in call to strings.MinRunes: non-concrete value string: