--- a/traverse.cue
+++ b/traverse.cue
@@ -27,7 +27,7 @@ package thema
 	inst: {...} // TODO consistently rename to 'object' or something
 
 	out: #SyntacticVersion
-	out: [ for _, sch in lin._sortedSchemas if ((sch._#schema & inst) != _|_) {sch.version}][0]
+	out: [ for _, sch in lin.schemas if ((sch._#schema & inst) != _|_) {sch.version}][0]
 }
 
 // #LinkedInstance represents data that is an instance of some schema, the