In #82, i added some native CUE logic to ensure that lenses and schemas were in a particular, expected sort order, regardless of the order the user provides them. This was conventional thinking about optimization - Translate operations needed to perform a lot of walking across these structures, and knowing that schemas and lenses were pre-sorted prior to the operation allowed retrieval of these objects by indexing, rather than more expensive search.

Unfortunately, the native CUE sorting - which relies on injected syntactic version comparators - is highly inefficient. And it's being performed pretty much everywhere that `#Lineage` is referenced and unified. Removing that sort logic resulted in almost absurd performance improvements...which is a good thing, because performance was categorically unacceptable before this change. Here's the [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) summary:

```
goos: darwin
goarch: amd64
pkg: github.com/grafana/thema
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
                                                                    │    old.txt    │               new.txt               │
                                                                    │    sec/op     │   sec/op     vs base                │
BasicTranslate/lineage/basic-multiversion/forward/simple-8            596.06m ± 87%   20.01m ± 7%  -96.64% (p=0.000 n=10)
BasicTranslate/lineage/basic-multiversion/reverse/withoutOptional-8   566.09m ± 22%   26.76m ± 4%  -95.27% (p=0.000 n=10)
BasicTranslate/lineage/basic-multiversion/reverse/withOptional-8      606.77m ± 81%   26.44m ± 4%  -95.64% (p=0.000 n=10)
UnifyLineage-8                                                        10.899m ±  9%   2.344m ± 4%  -78.50% (p=0.000 n=10)
BindLineage/PreUnified-8                                              17.367m ±  7%   4.262m ± 4%  -75.46% (p=0.000 n=10)
BindLineage/NotUnified-8                                              6784.9µ ± 10%   992.8µ ± 5%  -85.37% (p=0.000 n=10)
geomean                                                                80.04m         7.209m       -90.99%

                                                                    │    old.txt     │               new.txt                │
                                                                    │      B/op      │     B/op      vs base                │
BasicTranslate/lineage/basic-multiversion/forward/simple-8            72.680Mi ± 84%   3.399Mi ± 5%  -95.32% (p=0.000 n=10)
BasicTranslate/lineage/basic-multiversion/reverse/withoutOptional-8   69.136Mi ± 22%   3.944Mi ± 3%  -94.30% (p=0.000 n=10)
BasicTranslate/lineage/basic-multiversion/reverse/withOptional-8      74.231Mi ± 82%   3.928Mi ± 4%  -94.71% (p=0.000 n=10)
UnifyLineage-8                                                        1392.1Ki ±  0%   431.6Ki ± 0%  -69.00% (p=0.000 n=10)
BindLineage/PreUnified-8                                              2283.7Ki ±  0%   701.7Ki ± 0%  -69.27% (p=0.000 n=10)
BindLineage/NotUnified-8                                               808.7Ki ±  0%   187.0Ki ± 0%  -76.88% (p=0.000 n=10)
geomean                                                                9.813Mi         1.186Mi       -87.92%

                                                                    │    old.txt     │               new.txt               │
                                                                    │   allocs/op    │  allocs/op   vs base                │
BasicTranslate/lineage/basic-multiversion/forward/simple-8            1648.94k ± 84%   72.52k ± 5%  -95.60% (p=0.000 n=10)
BasicTranslate/lineage/basic-multiversion/reverse/withoutOptional-8   1567.07k ± 22%   83.76k ± 4%  -94.66% (p=0.000 n=10)
BasicTranslate/lineage/basic-multiversion/reverse/withOptional-8      1682.77k ± 82%   83.43k ± 4%  -95.04% (p=0.000 n=10)
UnifyLineage-8                                                         31.883k ±  0%   8.623k ± 0%  -72.95% (p=0.000 n=10)
BindLineage/PreUnified-8                                                57.47k ±  0%   15.91k ± 0%  -72.31% (p=0.000 n=10)
BindLineage/NotUnified-8                                               18.061k ±  0%   3.687k ± 0%  -79.59% (p=0.000 n=10)
geomean                                                                 228.9k         25.20k       -88.99%
```

The _smallest_ impact on our test suite was cutting execution time by more than two-thirds. For our worst operation, translate, execution time was reduced by more than 95%.


The tradeoff, of course, is that it is no longer acceptable for the user to write their `schemas` and `lenses` arrays in any order. I'm not sure that's much of a tradeoff though, TBH - i see little value in allowing flexibility in the order of declaration, anyway.

As-written, this PR is a UX regression in using Thema from pure CUE, as at least lenses will allow incorrect input orderings and only exhibit bizarre problems (arising from unmet lens index expectations) when running translate operations.

When using Thema from Go, this needn't be an issue, as we can enforce that the declared input ordering is correct in `BindLineage()`. We may be able to add some native CUE checks that enforce this, too, without hurting performance (no sorting, just asserting!). This PR doesn't add those checks, as i think we should defer that to a follow-up in order to stop the bleeding here as quickly as possible.

Fixes #137